# Compute style changes before update