@Kerumen 2022-01-21
@piaskowyk Do you plan to release a new version with this patch included? Thanks!